attempt fuseFormatResponse solution
